platform_in/requirements.txt

wheel
kafka-python
requests
elastic-apm
Flask
Flask-RESTful
Flask-Security
gunicorn
Jinja2
pytest
requests
certifi
pyrsistent
pydantic
sentry-sdk[flask]